# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для работы с фильмами и актерами в FilmHub. login: admin, password: kek",
        "title": "FilmHub API",
        "contact": {
            "url": "http://www.github.com/avealice"
        }
    },
    "host": "127.0.0.1:8000",
    "basePath": "/",
    "paths": {
        "/api/actor": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает нового актера.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/actor"
                ],
                "summary": "Создать актера.",
                "parameters": [
                    {
                        "description": "Данные нового актера",
                        "name": "actor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputActor"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Актер успешно создан",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Некорректный метод",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/actors": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получить всех актеров из базы данных.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/actors"
                ],
                "summary": "Получить всех актеров.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ActorWithMovies"
                            }
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Некорректный метод",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/movie": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новый фильм.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/movie"
                ],
                "summary": "Создать фильм",
                "parameters": [
                    {
                        "description": "Данные нового фильма",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputMovie"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Фильм создан успешно"
                    },
                    "400": {
                        "description": "Некорректный запрос или данные",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Некорректная роль",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/movie/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Выполняет поиск фильмов по указанным критериям (название или актер).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/movie/search"
                ],
                "summary": "Поиск фильмов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название фильма для поиска",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Имя актера для поиска",
                        "name": "actor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список фильмов, удовлетворяющих критериям поиска",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.MovieWithActors"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос или данные",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Некорректная роль",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/movies": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получает список всех фильмов с возможностью сортировки.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/movies"
                ],
                "summary": "Получить все фильмы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Критерий сортировки (title, rating, release_date)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Порядок сортировки (asc, desc)",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список фильмов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Movie"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос или данные",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Некорректный метод",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/sign-in": {
            "post": {
                "description": "Авторизует пользователя с заданными учетными данными и возвращает токен доступа",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/auth/"
                ],
                "summary": "Авторизация пользователя",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SignInInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Токен доступа",
                        "schema": {
                            "$ref": "#/definitions/handler.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос или данные",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/sign-up": {
            "post": {
                "description": "Регистрирует нового пользователя с заданными данными",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/auth/"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Данные нового пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID нового пользователя",
                        "schema": {
                            "$ref": "#/definitions/handler.UserIDResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос или данные",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ErrorResponse": {
            "description": "JSON-структура ответа с сообщением об ошибке.",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.SignInInput": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "handler.TokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "handler.UserIDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "model.Actor": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "description": "Birth date of the actor",
                    "type": "string"
                },
                "gender": {
                    "description": "Gender of the actor",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the actor",
                    "type": "string"
                }
            }
        },
        "model.ActorWithMovies": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "description": "Birth date of the actor",
                    "type": "string"
                },
                "gender": {
                    "description": "Gender of the actor",
                    "type": "string"
                },
                "id": {
                    "description": "Unique identifier for the actor",
                    "type": "integer"
                },
                "movies": {
                    "description": "Movies associated with the actor",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Movie"
                    }
                },
                "name": {
                    "description": "Name of the actor",
                    "type": "string"
                }
            }
        },
        "model.InputActor": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "description": "Birth date of the actor",
                    "type": "string"
                },
                "gender": {
                    "description": "Gender of the actor",
                    "type": "string"
                },
                "movies": {
                    "description": "Movies associated with the actor",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Movie"
                    }
                },
                "name": {
                    "description": "Name of the actor",
                    "type": "string"
                }
            }
        },
        "model.InputMovie": {
            "type": "object",
            "properties": {
                "actors": {
                    "description": "Actors associated with the movie",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Actor"
                    }
                },
                "description": {
                    "description": "Description of the movie",
                    "type": "string"
                },
                "rating": {
                    "description": "Rating of the movie",
                    "type": "integer"
                },
                "release_date": {
                    "description": "Release date of the movie",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the movie",
                    "type": "string"
                }
            }
        },
        "model.Movie": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description of the movie",
                    "type": "string"
                },
                "rating": {
                    "description": "Rating of the movie",
                    "type": "integer"
                },
                "release_date": {
                    "description": "Release date of the movie",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the movie",
                    "type": "string"
                }
            }
        },
        "model.MovieWithActors": {
            "type": "object",
            "properties": {
                "actors": {
                    "description": "Actors associated with the movie",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Actor"
                    }
                },
                "description": {
                    "description": "Description of the movie",
                    "type": "string"
                },
                "id": {
                    "description": "Unique identifier for the movie",
                    "type": "integer"
                },
                "rating": {
                    "description": "Rating of the movie",
                    "type": "integer"
                },
                "release_date": {
                    "description": "Release date of the movie",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the movie",
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "Password hash of the user",
                    "type": "string"
                },
                "username": {
                    "description": "Username of the user",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}